fix: 修复角色验证中类型转换问题
在validate_roles函数中，将角色ID显式转换为i64类型以避免潜在的类型错误

+++ sleepplanet-backend/src/routes/sys_admin/mod.rs
@@ -281,7 +281,7 @@ async fn validate_roles(role_ids: &[i64]) -> Result<Vec<i64>, AppError> {
     .await
     .map_err(|e| AppError::Internal(format!("角色验证失败: {}", e)))?;
 
-    Ok(roles.iter().map(|r| r.id).collect())
+    Ok(roles.iter().map(|r| r.id as i64).collect())
 }
 
 /// 获取用户角色名称
